moved APIs and common to utils instead of inside sys
