+++ astra-sim/network_frontend/ns3/AstraSimNetwork.cc
@@ -1,5 +1,5 @@
 #include <json/json.hpp>
-#include "astra-sim/system/AstraNetworkAPI.hh"
+#include "astra-sim/utils/AstraNetworkAPI.hh"
 #include "astra-sim/system/Sys.hh"
 #include "extern/remote_memory_backend/analytical/AnalyticalRemoteMemory.hh"
