+++ astra-sim/system/CSVWriter.cc
@@ -5,7 +5,7 @@ LICENSE file in the root directory of this source tree.
 
 #include "astra-sim/system/CSVWriter.hh"
 
-#include "astra-sim/system/Common.hh"
+#include "astra-sim/utils/Common.hh"
 
 #include <fcntl.h>
 #include <unistd.h>
