+++ astra-sim/system/Callable.hh
@@ -7,7 +7,7 @@ LICENSE file in the root directory of this source tree.
 #define __CALLABLE_HH__
 
 #include "astra-sim/system/CallData.hh"
-#include "astra-sim/system/Common.hh"
+#include "astra-sim/utils/Common.hh"
 
 namespace AstraSim {
