+++ astra-sim/system/MemBus.hh
@@ -9,7 +9,7 @@ LICENSE file in the root directory of this source tree.
 #include <string>
 
 #include "astra-sim/system/Callable.hh"
-#include "astra-sim/system/Common.hh"
+#include "astra-sim/utils/Common.hh"
 
 namespace AstraSim {
