+++ astra-sim/system/MemMovRequest.hh
@@ -7,7 +7,7 @@ LICENSE file in the root directory of this source tree.
 #define __MEM_MOV_REQUEST_HH__
 
 #include "astra-sim/system/Callable.hh"
-#include "astra-sim/system/Common.hh"
+#include "astra-sim/utils/Common.hh"
 #include "astra-sim/system/SharedBusStat.hh"
 #include "astra-sim/system/Sys.hh"
