+++ astra-sim/system/Sys.hh
@@ -8,8 +8,8 @@ LICENSE file in the root directory of this source tree.
 
 #include <chrono>
 
-#include "astra-sim/system/AstraNetworkAPI.hh"
-#include "astra-sim/system/AstraRemoteMemoryAPI.hh"
+#include "astra-sim/utils/AstraNetworkAPI.hh"
+#include "astra-sim/utils/AstraRemoteMemoryAPI.hh"
 #include "astra-sim/system/Callable.hh"
 #include "astra-sim/system/CollectivePhase.hh"
 #include "astra-sim/system/CommunicatorGroup.hh"
