+++ astra-sim/system/UsageTracker.hh
@@ -10,7 +10,7 @@ LICENSE file in the root directory of this source tree.
 
 #include "astra-sim/system/CSVWriter.hh"
 #include "astra-sim/system/Callable.hh"
-#include "astra-sim/system/Common.hh"
+#include "astra-sim/utils/Common.hh"
 #include "astra-sim/system/Usage.hh"
 
 namespace AstraSim {
