+++ astra-sim/utils/Common.hh
